Refactor all game-related methods into GameBoard.java. Fix bug that caused winning on the last move to result in a 'TIE'
