+++ src/GUI.java
@@ -2,60 +2,19 @@ import javax.swing.*;
 import javax.swing.border.LineBorder;
 import java.awt.*;
 import java.awt.event.ActionListener;
-import java.util.HashMap;
-import java.util.Random;
 
 public class GUI {
 
     private static JPanel tttBoard;
     private static GameBoard board;
-    private static HashMap componentMap;
+
 
     public static void main(String[] args) {
 
         board = new GameBoard();
         makeGameBoard();
-        createComponentMap();
+        GameBoard.createComponentMap(tttBoard);
 
-    }
-
-    private static void computerTurn() {
-        Random rNum = new Random();
-        String cpuChoice = Integer.toString(rNum.nextInt(8));
-        Component findButton = getComponentByName(cpuChoice);
-        if (findButton == null) {
-            throw new AssertionError();
-        }
-        String text = ((JButton)findButton).getText();
-
-        while (!checkSelected(text)) {
-            cpuChoice = Integer.toString(rNum.nextInt(8));
-            findButton = getComponentByName(cpuChoice);
-            assert findButton != null;
-            text = ((JButton)findButton).getText();
-        }
-        ((JButton)findButton).setText("O");
-        findButton.setForeground(Color.ORANGE);
-        board.addCpuChoice(Integer.parseInt(cpuChoice));
-        String result = board.checkWinner();
-        if (result.length() > 0) {
-            System.out.println(result);
-        }
-    }
-
-    private static void createComponentMap() {
-        componentMap = new HashMap<String, JButton>();
-        Component[] buttons = tttBoard.getComponents();
-        for (Component button : buttons) {
-            componentMap.put(button.getName(), button);
-        }
-    }
-
-    private static Component getComponentByName(String name) {
-        if (componentMap.containsKey(name)) {
-            return (Component) componentMap.get(name);
-        }
-        else return null;
     }
 
     public static void makeGameBoard() {
@@ -91,13 +50,11 @@ public class GUI {
         frame.setVisible(true);
     }
 
-    private static boolean checkSelected (String bTxt) {
-        return (!bTxt.contains("X") && !bTxt.contains("O"));
-    }
+
 
     private static ActionListener selectMove(JButton b) {
         return e -> {
-            if (checkSelected(b.getText())) {
+            if (board.checkSelected(b.getText())) {
                 b.setText("X");
                 b.setForeground(Color.CYAN);
                 board.addUserChoice(Integer.parseInt(b.getName()));
@@ -105,7 +62,7 @@ public class GUI {
                 if (result.length() > 0) {
                     System.out.println(result);
                 } else {
-                    computerTurn();
+                    GameBoard.computerTurn(board);
                 }
             }
         };
